+++ database/dataSource.ts
@@ -113,6 +113,7 @@ import { AjoutBudgetEtFinanceAProfilETSanitaire1797688226682 } from "./migration
 import { AjoutAllocationDeRessourcesToBudgetEtFinance1798688226682 } from "./migrations/1798688226682-AjoutAllocationDeRessourcesToBudgetEtFinance";
 import { AjoutDesOccupationsDansLesProfils1799478704013 } from "./migrations/1799478704013-AjoutDesOccupationsDansLesProfils";
 import { AjoutLesJourneesUsldDansLesProfils1799501916707 } from "./migrations/1799501916707-AjoutLesJourneesUsldDansLesProfils";
+import { AjoutLienHypertexteMessageAccueil1806000000000 } from "./migrations/1806000000000-AjoutLienHypertexteMessageAccueil";
 import { AjoutVigieRhContrat1738320902145 } from "./migrations/vigie_rh/1738320902145-AjoutVigieRhContrat";
 import { AjoutVigieRhProfessionFiliere1738328763147 } from "./migrations/vigie_rh/1738328763147-AjoutVigieRhProfessionFiliere";
 import { AjoutVigieRhProfessionGroupe1738328796095 } from "./migrations/vigie_rh/1738328796095-AjoutVigieRhProfessionGroupe";
@@ -411,7 +412,8 @@ const datasource = new DataSource({
     AjoutProfilTresorerie1773223041788,
     AjoutColonnesAutorisationsAmmStatutEtLibelles1774365753368,
     SuppressionNomenclatureAMM1774521749651,
-    AjoutTableMessageAccueil1779354621997
+    AjoutTableMessageAccueil1779354621997,
+    AjoutLienHypertexteMessageAccueil1806000000000,
   ],
   type: "postgres",
   url: environmentVariables.DATABASE_URL,